backend/document-ingestion/requirements.txt
# Document Ingestion Service Dependencies

# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# File handling and validation
python-magic==0.4.27
Pillow==10.1.0

# Data validation
pydantic==2.5.0

# Utilities
python-dotenv==1.0.0
pathlib2==2.3.7

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
flake8==6.1.0
isort==5.12.0
